fix: Widen PWM duty API to 32-bit values

+++ Platform/Inc/platform_pwm.h
@@ -12,6 +12,6 @@
 Platform_StatusTypeDef Platform_PWM_Start(Platform_TIM_Instance tim, Platform_TIM_Channel channel);
 Platform_StatusTypeDef Platform_PWM_Stop(Platform_TIM_Instance tim, Platform_TIM_Channel channel);
 void Platform_PWM_SetCompare(Platform_TIM_Instance tim, Platform_TIM_Channel channel, uint32_t compare);
-void Platform_PWM_SetDuty(Platform_TIM_Instance tim, Platform_TIM_Channel channel, uint16_t max_duty, uint16_t duty);
+void Platform_PWM_SetDuty(Platform_TIM_Instance tim, Platform_TIM_Channel channel, uint32_t max_duty, uint32_t duty);
 
 #endif //LAYERED_EMBEDDED_C_PLATFORM_PWM_H
